Removed timerit and dont expose util_deprecated
