+++ ubelt/__init__.py
@@ -28,43 +28,40 @@ Testing:
 
 __version__ = '1.0.0'
 
-__submodules__ = [
-    'util_arg',
-    'util_cache',
-    'util_colors',
-    'util_const',
-    'util_cmd',
-    'util_dict',
-    'util_download',
-    'util_download_manager',
-    'util_func',
-    'util_format',
-    'util_futures',
-    'util_io',
-    'util_links',
-    'util_list',
-    'util_hash',
-    'util_import',
-    'util_indexable',
-    'util_memoize',
-    'util_mixins',
-    'util_path',
-    'util_platform',
-    'util_str',
-    'util_stream',
-    'util_time',
-    'util_zip',
-    'orderedset',
-    'progiter',
-    'timerit',
-    '_util_deprecated',
-]
+__submodules__ = {
+    'util_arg': None,
+    'util_cache': None,
+    'util_colors': None,
+    'util_const': None,
+    'util_cmd': None,
+    'util_dict': None,
+    'util_download': None,
+    'util_download_manager': None,
+    'util_func': None,
+    'util_format': None,
+    'util_futures': None,
+    'util_io': None,
+    'util_links': None,
+    'util_list': None,
+    'util_hash': None,
+    'util_import': None,
+    'util_indexable': None,
+    'util_memoize': None,
+    'util_mixins': None,
+    'util_path': None,
+    'util_platform': None,
+    'util_str': None,
+    'util_stream': None,
+    'util_time': None,
+    'util_zip': None,
+    'orderedset': None,
+    'progiter': None,
+    # '_util_deprecated': []
+}
 
 # <AUTOGEN_INIT>
-from ubelt import _util_deprecated
 from ubelt import orderedset
 from ubelt import progiter
-from ubelt import timerit
 from ubelt import util_arg
 from ubelt import util_cache
 from ubelt import util_cmd
@@ -129,23 +126,21 @@ from ubelt.util_platform import (DARWIN, LINUX, POSIX, WIN32,
 from ubelt.util_str import (codeblock, ensure_unicode, hzcat, indent,
                             paragraph,)
 from ubelt.util_stream import (CaptureStdout, CaptureStream, TeeStringIO,)
-from ubelt.util_time import (timestamp,)
+from ubelt.util_time import (Timer, timestamp,)
 from ubelt.util_zip import (split_archive, zopen,)
 from ubelt.orderedset import (OrderedSet, oset,)
 from ubelt.progiter import (ProgIter,)
-from ubelt.timerit import (Timer, Timerit,)
 
 __all__ = ['AutoDict', 'AutoOrderedDict', 'CacheStamp', 'Cacher',
            'CaptureStdout', 'CaptureStream', 'DARWIN', 'DownloadManager',
            'Executor', 'FormatterExtensions', 'IndexableWalker', 'JobPool',
            'LINUX', 'NO_COLOR', 'NiceRepr', 'NoParam', 'OrderedSet', 'POSIX',
-           'Path', 'ProgIter', 'TeeStringIO', 'TempDir', 'Timer', 'Timerit',
-           'WIN32', '_util_deprecated', 'allsame', 'argflag', 'argmax',
-           'argmin', 'argsort', 'argunique', 'argval', 'augpath', 'boolmask',
-           'chunks', 'cmd', 'codeblock', 'color_text', 'compatible',
-           'compress', 'ddict', 'delete', 'dict_diff', 'dict_hist',
-           'dict_isect', 'dict_subset', 'dict_union', 'download', 'dzip',
-           'ensure_app_cache_dir', 'ensure_app_config_dir',
+           'Path', 'ProgIter', 'TeeStringIO', 'TempDir', 'Timer', 'WIN32',
+           'allsame', 'argflag', 'argmax', 'argmin', 'argsort', 'argunique',
+           'argval', 'augpath', 'boolmask', 'chunks', 'cmd', 'codeblock',
+           'color_text', 'compatible', 'compress', 'ddict', 'delete',
+           'dict_diff', 'dict_hist', 'dict_isect', 'dict_subset', 'dict_union',
+           'download', 'dzip', 'ensure_app_cache_dir', 'ensure_app_config_dir',
            'ensure_app_data_dir', 'ensure_unicode', 'ensuredir', 'expandpath',
            'find_duplicates', 'find_exe', 'find_path', 'flatten',
            'get_app_cache_dir', 'get_app_config_dir', 'get_app_data_dir',
@@ -159,9 +154,9 @@ __all__ = ['AutoDict', 'AutoOrderedDict', 'CacheStamp', 'Cacher',
            'platform_cache_dir', 'platform_config_dir', 'platform_data_dir',
            'progiter', 'readfrom', 'repr2', 'shrinkuser', 'sorted_keys',
            'sorted_vals', 'split_archive', 'split_modpath', 'symlink', 'take',
-           'timerit', 'timestamp', 'touch', 'unique', 'unique_flags',
-           'userhome', 'util_arg', 'util_cache', 'util_cmd', 'util_colors',
-           'util_const', 'util_dict', 'util_download', 'util_download_manager',
+           'timestamp', 'touch', 'unique', 'unique_flags', 'userhome',
+           'util_arg', 'util_cache', 'util_cmd', 'util_colors', 'util_const',
+           'util_dict', 'util_download', 'util_download_manager',
            'util_format', 'util_func', 'util_futures', 'util_hash',
            'util_import', 'util_indexable', 'util_io', 'util_links',
            'util_list', 'util_memoize', 'util_mixins', 'util_path',
